Filter REFERENCEFRAME and PIXEL
For some reason, the equation solver is returning all
the equations for frame and pixel, most likely caused
by the fact that there are no equations for either of
these in the system. For the moment, we simply remove
them from the print out. A better method might remove
all units which are missing from the input or test to
see if the unit is in the resulting equation(s). This
*does* check to see if any inappropriate symbol is in
the output.